Use FIFREEZE instead of F2FS_IOC_SET_PIN_FILE
TestF2fsCompression is flaky, with a *very* low failure rate.  I believe
the problem is that f2fs is moving the blocks of the file after their
locations have been queried with FS_IOC_FIEMAP, but before they have
been read from the block device.  F2FS_IOC_SET_PIN_FILE was meant to
prevent this, but actually it's unsupported on compressed files.

Instead, switch to the FIFREEZE ioctl, which "freezes" the filesystem.
Anyone writing to the filesystem is forced to wait until FITHAW.

It's not great that we've ended up here, as freezing the /data
filesystem isn't something that Android ever does; there's a chance that
it will cause problems.  But I haven't seen any problems when running
the test repeatedly.

Test: atest vts_kernel_encryption_test # Cuttlefish, f2fs
Test: atest vts_kernel_encryption_test # Cuttlefish, ext4
Bug: 205123902
Change-Id: I0da1a6dfa6f3091154c1b2e61771adf97ee943ba