Make some generic ramdisk items optional for S devices
In T, we added snapuserd to generic ramdisk. Snapuserd is required to be
present in init_boot. But for S devices(which do not have init_boot),
snapuserd is optional. To accomodate this, guard T only requirements
with ro.bootimage.build.version.sdk .

AOSP link:
https://android-review.googlesource.com/c/platform/test/vts-testcase/kernel/+/2118667/

This CL needs to be in tm-dev because without it, VTS tests on S devices
or devices running T system + S vendor would fail.

Bug: 233159800
Change-Id: I05a8d99cdd0eb924421f3f3e13472aaee7ef4c2c