Skip GenericBootImageTest on 32-bit ARM devices.
GKI and generic boot image requirements are not
enforced on 32-bit ARM devices.

Keep the test running on x86 and x86_64 even
though the requirement is not in place so that
the test is kept green on virtual devices.

Test: builds
Bug: 237022520
Change-Id: I7f8dfb5d5c7f32129c350fec3f578f668fb166a7
(cherry picked from commit 3f8098a2b55289e432901679a4a532834a4b6a03)