16k: VTS: Check vendor ELFs are 64KB aligned
Check that ELF files on the vendor partition are at least
64KB aligned starting from U (34). Android makes sparse copies
of the files so the *real* disk space usage is the same as 4KB
aligned binaries. Note tools like `ls` will still show an
increase in disk space usage, but that is only the apparent file
size, as the same 4KB block size is used in the read only
partitions.

Bug: 277360995
Test: atest -c vendor_elf_alignment_test -s <serial>
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>
Change-Id: Id1da3c44391f6ebc42c0d94e7f11424a90674234