isa: State the Requirement ID related to testing for 64-bit kernel usage
The requirement of using a 64-bit kernel for devices launching with
Android 14 and later versions follows from VSR-3.12, which states that
devices launching with Android 14 must support 64-bit NDK ABIs, which
is only possible if a 64-bit kernel is in use. This is currently not
clear from the existing error message that is emitted in case the test
fails, so cite the VSR requirement ID for reference.

Bug: 253608934
Bug: 279977975
Change-Id: I252cab2c0278bc03141316c9b2176daf7d57857f
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>