vts_kernel_encryption_test: Retry FIFREEZE if it fails with EINVAL
Very rarely, on f2fs the FIFREEZE ioctl fails with EINVAL.
Unfortunately, the reason for this is still unknown.  While this is a
kernel bug, and the point of vts_kernel_encryption_test is to detect
bugs, it's not really the type of bug the test is looking for.  We can't
have the test be forever flaky because of this.

Based on preliminary analysis, there's a good chance that this is caused
by a race condition with certain types of f2fs filesystem modifications.
So, there's a good chance that retrying the ioctl will help.  Do this,
and also bump the timeout up to 20 seconds for good measure.

Bug: 255800104
Bug: 287002682
Test: atest vts_kernel_encryption_test
Change-Id: I9dc627fd7ae1b18f7ccb8a0286ef2305d1a87ca9