WrappedKey: Support alternate KDF context for HW-wrapped keys
Add alternative context to allow key derivation from trusted
execution environment (TEE) as per the hardware requirement.

Background:
Some legacy SOCs, key derivation is tied to execution
environment for more security.
key derivation needs to be in TEE for virtualization
environment cases which are common for Auto.
It is different to existing scenarios where key derivation
is tied only to HLOS.

Hence the policy for key derivation needs modification
accordingly.

In summary, it is recommended to use existing KDF context
for all the latest SoCs and even for legacy SoCs with kdf
not-tied to TEE.

Implementation Details:
Use vendor property to decide which context to be used.
All required tests will use only one context, based on
vendor property.

Bug: 287012296
Test: All VTS kernel encryption tests are passed with new
      and existing context.

Change-Id: If7ae3f4ff5efeeb467d8ac2cbd604a69267369dc
Signed-off-by: Santosh Dronamraju <quic_sdronamr@quicinc.com>