BPF: remove duplicate cflags and update btf flags.
With the addition of the -Wall, -Werror, -Wextra cflags to the bpf soong
module as default, they are no longer necessary in the bpf definition in
Android.bp files.

Additionally, 'btf: true'  is now set by default by the build system. To counter this and maintain the existing behavior, add 'btf: false' where necessary.

Test: treehugger
Change-Id: Ieaa1d4c3b128ee78d7eb4845e89f8afb9ae3dc03
Signed-off-by: Neill Kapron <nkapron@google.com>