[VTS] Support custom FBE derivation algos
For hardware compatability, adds support for a custom KDF implementation
that processes the fixed-input string from back-to-front. Uses an
Android property to select the KDf as well as the context/label.

Additionally adds a test to validate that the outputs of each of the KDF
strategies is consistent.

Bug: 353759470
Bug: 330405838
Flag: TEST_ONLY
Test: vts_kernel_encryption_test

Change-Id: I71bdda56988e1f726b09c2648f73c8d43c0792b0
Signed-off-by: Ross Kettleson <kettro@google.com>