Allow OEMs to update old 32/64 devices' vendor partitions.
Without this change, a 32/64 device would suddenly hit the "you need to be 32-only or 64-only" restriction if the vendor partition was updated past API 34. Obviously the intention here is to prevent _new_ 32/64 devices, not prevent OEMs from keeping old devices updated for longer.

Add an early out for this case.

Bug: https://issuetracker.google.com/321167077
Change-Id: Ieca1d6436814f786d555ff1fa2d38881be433511