Test file-based encryption with a sub-block data unit size
Add tests for encryption policies that specify a data unit size of 4K.
Previously only the default data unit size, which is equal to the
filesystem block size, was tested.  This matters when the filesystem
block size is not 4K.

This provides test coverage for the dusize_4k encryption setting, which
was added in order to allow devices to use a 16K block size filesystem
with inline encryption hardware that only supports 4K data units.

Bug: 337068060
Test: 'atest vts_kernel_encryption_test', with 4K and 16K block sizes.
Change-Id: I38db7a410d4bae190584082609a7854cd519860c