Disable `LTP kernel_misc.cpufreq_boost` tests.
The cpufreq_boost tests will often fail as when they are run previous
tests cause the device to be warm, and thus there isn't the thermal
availability for boosting. This caused the test to fail.

Pull it out of the stable set for now and add it to the disabled tests.

Bug: 397429079
Bug: 446729350
Flag: EXEMPT BUGFIX
Change-Id: I89f69c2097c28dba817b77c510efecbe71e70c61
Signed-off-by: John Stultz <jstultz@google.com>