Test wrappedkey encryption option
For each existing hardware-wrapped key test case, add a version of the
test case that uses the modern-style hardware-wrapped keys instead.
This means using the blk-crypto ioctls to import and prepare the keys,
and using the new fscrypt policy flag.  This corresponds to the fstab
encryption option "wrappedkey", vs. the existing "wrappedkey_v0".

Note that as with the "wrappedkey_v0" test cases, the "wrappedkey" test
cases are skipped on devices that don't support them.

Bug: 420692910
Test: atest vts_kernel_encryption_test # on several different devices
Flag: EXEMPT TEST_ONLY
Change-Id: Ib06a15f96b89c698600d195048b7dda04b7c23c6