Also accept EINVAL for blk-crypto ioctls unsupported
When the block device belongs to a SCSI disk such as a UFS device and
the kernel is too old to know about the blk-crypto ioctls, the error
code is EINVAL instead of the expected ENOTTY.

Unfortunately, this seems to be a longstanding bug in the kernel's SCSI
driver, and it's unclear whether it can even be fixed at this point.

For now just make the test accept EINVAL too.

Bug: 420692910
Bug: 455989000
Test: atest vts_kernel_encryption_test # on device w/ UFS and 6.6 kernel
Flag: EXEMPT TEST_ONLY
Change-Id: Ie3732054d90aa308237ae504e51d420dacfae781