Pass correct raw block device down to ImportAndPrepareHwWrappedKey()
As mentioned in the comment for GetRawBlockDevice(),
FstabEntry::blk_device can't be used because it sometimes contains a
logical partition name rather than a real block device.  Specifically
this happens when the system has been booted as a "Dynamic System
Update".  Therefore, GetRawBlockDevice() walks backwards from the
filesystem to get the correct device.  I missed that this is needed, and
ImportAndPrepareHwWrappedKey() circumvents that logic and instead uses
the FstabEntry.  Pass the correct raw block device down to it instead.

This fixes test failures with the message:

    Failed to open userdata_gsi: No such file or directory

Bug: 458744716
Test: atest vts_kernel_encryption_test
Flag: EXEMPT TEST_ONLY
Change-Id: I721376ef287538505c487cff949d0c32080e7705